+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [v1.2.0] - March 2026
+
+* Fix -c and -l not working
+* Fix service launch failture due to privilege issue
+* Separate log files for hotkeys service and agent (--log now expects a directory)
+
 ## [v1.1.0] - 2 February 2026
 
 Install and run as service